Refactor OrgPageMap by extracting functions and moving files
