+++ src/features/organizations/utils/markerImage.ts
